Reorganize project files
