+++ src/app/models/index.ts
@@ -1,0 +1,2 @@
+export * from './page.model';
+export * from './product.model';
